Processor time measurement included, one file option checked Bonus: tester code and sample .csv

+++ utilL.py
@@ -23,18 +23,30 @@ def createFile():
     concName = os.path.join(configL.filePath, firstPart + numberExp +'.csv')
     print concName
     
-    while True == os.path.isfile(concName):
-        modL.increaseVal()
-        numberExp = '-' + str(configL.counter)
-        concName = os.path.join(configL.filePath, firstPart + numberExp +'.csv')
-    
+    #this section gives the opportunity for saving measurement into separate files
+    # while True == os.path.isfile(concName):
+    #     modL.increaseVal()
+    #     numberExp = '-' + str(configL.counter)
+    #     concName = os.path.join(configL.filePath, firstPart + numberExp +'.csv')
+
+    if True != os.path.isfile(concName):
+        fileL = open(concName, 'w')
+        fileL.close()
+
     modL.updateExactPath(concName)
-    fileL = open(concName, 'w')
-    fileL.write(str(firstPart) + ' No# '+ str(configL.counter) + '\n')
+
+    noOfExp = modL.quantityCount()
+    
+    fileL = open(concName, 'a')
+    
+    if noOfExp != 1:
+        fileL.write('' + '\n')
+
+    fileL.write(str(firstPart) + ' No# '+ str(noOfExp) + '\n')
     fileL.write('Sent signal: ' + str(configL.mSent) + '\n')
     fileL.write('zDistance: ' + str(configL.zDistance) + '\n')
     fileL.close()
-    writeLine('Cycles','Alpha','Ratio')
+    writeLine('Cycles','Alpha','Ratio','Time Difference')
 
     return configL.counter
 
@@ -45,13 +57,13 @@ def findMax(numpyArray):
 
     return np.amax(numpyArray, axis=None)
 
-def writeLine(label,alpha,ratio):
+def writeLine(label,alpha,ratio,timeDiff):
     """
     Append data to the file.
     """
     fileL = open(configL.exactPath, 'ab')
     toFile = csv.writer(fileL)
-    toFile.writerow([label,alpha,ratio])
+    toFile.writerow([label,alpha,ratio,timeDiff])
     fileL.close()
 
 def handleL(numpyArray):
@@ -59,6 +71,11 @@ def handleL(numpyArray):
     Wrap up.
     Assumption: the type of the argument numpyArray is numpy array.
     """
+
+    if configL.timeMeasure == 0:
+        modL.startTimer()
+    else:
+        modL.calcTime()
 
     if configL.exactPath == '':
         createFile()
@@ -72,5 +89,5 @@ def handleL(numpyArray):
     modL.updateAlphas(configL.alpha)
     modL.updateRatio(configL.alphas)
     alphaToWrite = configL.alphas[1]
-    writeLine(configL.label,alphaToWrite,configL.ratio)
+    writeLine(configL.label,alphaToWrite,configL.ratio, configL.timeDifference)
     modL.updateLabel()
